A new filter trait which searches for other jobs
If the same Git repository is configured in another multibranch pipeline
(MBP), then it will be considered a match and will allow creating the
same job in the current multibranch pipeline.

This filter is useful if you have multiple multibranch pipelines
configured for one Git repository and want to treat another MBP as the
"primary".  If a matching job doesn't exist in the primary MBP, then the
current pipeline configured with the JobExistsElsewhereTrait will not
create a matching job in this MBP.